Device Code Improvements and Shadow Update Lambda
Adding python code for a Lambda function the submit desired state changes to IoT Core, with filtering to ensure only valid and acceptable values are passed through.
Adding a "first boot" check to the device code to make boot process more robust, with regards to ensuring some URL is shared as soon as the device get's it's desired state from Iot Core.